refactor(ai-agent): move data app read types to backend, aggregate as static method
- DataAppRead* types out of common apps/code (not content-as-code) into AiAgentToolsService/dataAppRead
- aggregateDataAppDataReferences → AiAgentToolsService static, Set-based dedupe
- generic readContent description
- regenerate mcp-tools snapshot (CI)
