+++ packages/backend/src/ee/services/AiAgentToolsService/AiAgentToolsService.ts
@@ -41,10 +41,10 @@ import {
     type ChartAsCode,
     type CustomChartType,
     type DashboardAsCode,
-    type DataAppRead,
     type DataAppVizSchema,
     type FieldValueSearchResult,
     type ParameterDefinitions,
+    type PersistedDataAppDataReferences,
     type SchedulerAiAugmentation,
 } from '@lightdash/common';
 import * as JsonPatch from 'fast-json-patch';
@@ -147,7 +147,11 @@ import type {
 } from '../AppGenerateService/AppGenerateService';
 import { PreviewDeploySetupService } from '../PreviewDeploySetupService/PreviewDeploySetupService';
 import type { SchedulerAiAugmentationService } from '../SchedulerAiAugmentationService/SchedulerAiAugmentationService';
-import { aggregateDataAppDataReferences } from './aggregateDataAppDataReferences';
+import type {
+    DataAppRead,
+    DataAppReadDataReferences,
+    DataAppReadExploreReferences,
+} from './dataAppRead';
 
 type AgentListContentResult = Awaited<ReturnType<ListContentFn>>;
 type AgentListContentItem = AgentListContentResult['items'][number];
@@ -1720,6 +1724,109 @@ export class AiAgentToolsService extends BaseService {
         }
     }
 
+    // Per-call-site references → per-explore summary. Locations and custom SQL
+    // text are dropped; charts missing from `chartSlugsByUuid` (deleted) too.
+    static aggregateDataAppDataReferences(
+        { references, stats }: PersistedDataAppDataReferences,
+        chartSlugsByUuid: Readonly<Record<string, string>>,
+    ): DataAppReadDataReferences {
+        const pushUnique = (target: string[], values: string[]) => {
+            const seen = new Set(target);
+            for (const value of values) {
+                if (!seen.has(value)) {
+                    target.push(value);
+                    seen.add(value);
+                }
+            }
+        };
+        const explores = new Map<string, DataAppReadExploreReferences>();
+        const linkedCharts = new Map<string, string[]>();
+        const externalConnections = new Map<string, string[]>();
+        const unresolved = new Set<string>();
+
+        const exploreFor = (name: string) => {
+            const existing = explores.get(name);
+            if (existing) return existing;
+            const created: DataAppReadExploreReferences = {
+                name,
+                dimensions: [],
+                metrics: [],
+                filterFields: [],
+                sortFields: [],
+                parameterKeys: [],
+                localFields: [],
+                customSqlFieldCount: 0,
+            };
+            explores.set(name, created);
+            return created;
+        };
+
+        for (const ref of references) {
+            ref.unresolved.forEach((part) => unresolved.add(part));
+            switch (ref.kind) {
+                case 'query': {
+                    if (ref.explore === null) break;
+                    const explore = exploreFor(ref.explore);
+                    pushUnique(explore.dimensions, ref.dimensions);
+                    pushUnique(explore.metrics, ref.metrics);
+                    pushUnique(explore.filterFields, [
+                        ...ref.dimensionFilterFields,
+                        ...ref.metricFilterFields,
+                    ]);
+                    pushUnique(explore.sortFields, ref.sortFields);
+                    pushUnique(explore.parameterKeys, ref.parameterKeys);
+                    pushUnique(explore.localFields, ref.localFields);
+                    if (ref.customSql) {
+                        explore.customSqlFieldCount +=
+                            ref.customSql.tableCalculations.length +
+                            ref.customSql.customDimensions.length +
+                            ref.customSql.additionalMetrics.length;
+                    }
+                    break;
+                }
+                case 'globalFilter': {
+                    if (ref.explore === null) break;
+                    const fields = ref.fields ?? (ref.field ? [ref.field] : []);
+                    pushUnique(exploreFor(ref.explore).filterFields, fields);
+                    break;
+                }
+                case 'savedChart': {
+                    const slug =
+                        ref.chartUuid === null
+                            ? undefined
+                            : chartSlugsByUuid[ref.chartUuid];
+                    if (slug === undefined) break;
+                    const filterFields = linkedCharts.get(slug) ?? [];
+                    pushUnique(filterFields, ref.filterFields);
+                    linkedCharts.set(slug, filterFields);
+                    break;
+                }
+                case 'externalFetch': {
+                    if (ref.alias === null) break;
+                    const paths = externalConnections.get(ref.alias) ?? [];
+                    if (ref.path !== null) pushUnique(paths, [ref.path]);
+                    externalConnections.set(ref.alias, paths);
+                    break;
+                }
+                default:
+                    assertUnreachable(ref, 'Unknown data reference kind');
+            }
+        }
+
+        return {
+            explores: [...explores.values()],
+            linkedCharts: [...linkedCharts].map(([slug, filterFields]) => ({
+                slug,
+                filterFields,
+            })),
+            externalConnections: [...externalConnections].map(
+                ([alias, paths]) => ({ alias, paths }),
+            ),
+            stats,
+            unresolved: [...unresolved].sort(),
+        };
+    }
+
     private async buildDataAppRead(
         source: DataAppReadSource,
     ): Promise<DataAppRead> {
@@ -1788,7 +1895,7 @@ export class AiAgentToolsService extends BaseService {
                 ).map((connection) => connection.alias),
             },
             dataReferences: dataReferences
-                ? aggregateDataAppDataReferences(
+                ? AiAgentToolsService.aggregateDataAppDataReferences(
                       dataReferences,
                       chartSlugsByUuid,
                   )
