+++ packages/common/src/schemas/json/mcp-tools-1.0.json
@@ -374,7 +374,7 @@
                 "openWorldHint": false,
                 "readOnlyHint": true
             },
-            "description": "Tool: \"find_content\"\nPurpose:\nFinds spaces, charts, dashboards, or Data Apps by name or description within a project, returning detailed information about each.\n\nUsage tips:\n- IMPORTANT: Pass the user's full query or relevant portion directly (e.g., \"revenue based on campaigns\" instead of just \"campaigns\").\n- The search engine understands natural language and context — more descriptive queries yield better results.\n- You can provide multiple search queries to look for different topics simultaneously (e.g., [\"monthly revenue\", \"user acquisition trends\"]).\n- Pass spaceSlug to search only inside that space and its descendants.\n- verifiedOnly=true searches only admin-verified charts/dashboards (spaces omitted). Data Apps have no verification state and remain included. Verification is a status, not text — search topic keywords, never the word \"verified\". If nothing matches, verified content may still exist under other terms, so never claim the project has none from one query. Re-run with verifiedOnly=false only if unverified charts or dashboards are acceptable.\n- If results aren't relevant, retry with the full user query or more specific terms.\n- Dashboards with validation errors will be deprioritized.\n- Returns space breadcrumb/path metadata and canonical chart, dashboard, and Data App URLs when available.\n- Dashboards show a preview of the first 5 charts and the total chart count. Use \"read_content\" to see all charts for a specific dashboard.\n- It doesn't provide summaries for dashboards yet, so don't suggest this capability.",
+            "description": "Tool: \"find_content\"\nPurpose:\nFinds spaces, charts, dashboards, or Data Apps by name or description within a project, returning detailed information about each.\n\nUsage tips:\n- IMPORTANT: Pass the user's full query or relevant portion directly (e.g., \"revenue based on campaigns\" instead of just \"campaigns\").\n- The search engine understands natural language and context — more descriptive queries yield better results.\n- You can provide multiple search queries to look for different topics simultaneously (e.g., [\"monthly revenue\", \"user acquisition trends\"]).\n- Pass spaceSlug to search only inside that space and its descendants.\n- verifiedOnly=true searches only admin-verified charts/dashboards (spaces omitted). Data Apps have no verification state and remain included. Verification is a status, not text — search topic keywords, never the word \"verified\". If nothing matches, verified content may still exist under other terms, so never claim the project has none from one query. Re-run with verifiedOnly=false only if unverified charts or dashboards are acceptable.\n- If results aren't relevant, retry with the full user query or more specific terms.\n- Dashboards with validation errors will be deprioritized.\n- Returns space breadcrumb/path metadata and canonical chart, dashboard, and Data App URLs when available.\n- Dashboards show a preview of the first 5 charts and the total chart count. Use \"read_content\" to see all charts for a specific dashboard.\n- Data Apps are readable: call \"read_content\" with type \"data_app\" and the app slug to see what the app shows, the charts and dashboard it was generated from, and the explores, fields, filters, parameters and external connections it queries.\n- It doesn't provide summaries for dashboards yet, so don't suggest this capability.",
             "inputSchema": {
                 "$schema": "http://json-schema.org/draft-07/schema#",
                 "additionalProperties": false,
@@ -1899,19 +1899,19 @@
                 "openWorldHint": false,
                 "readOnlyHint": true
             },
-            "description": "Read a dashboard or chart as JSON using its slug. Call this before editing.",
+            "description": "Read a dashboard, chart, or data app as JSON using its slug. Call this before editing a dashboard or chart. Data apps are read-only and returned without code.",
             "inputSchema": {
                 "$schema": "http://json-schema.org/draft-07/schema#",
                 "additionalProperties": false,
                 "properties": {
                     "slug": {
-                        "description": "Slug of the dashboard or chart to read.",
+                        "description": "Slug of the dashboard, chart, or data app to read.",
                         "minLength": 1,
                         "type": "string"
                     },
                     "type": {
                         "description": "Type of Lightdash content to read.",
-                        "enum": ["dashboard", "chart"],
+                        "enum": ["dashboard", "chart", "data_app"],
                         "type": "string"
                     }
                 },
